backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
google-generativeai==0.8.3
python-multipart==0.0.9
pillow==10.4.0
httpx==0.27.2
pydantic==2.9.2
